refactor: [US-053] - Extract dna/reader_type.py; promote DIVERSITY_THRESHOLD/DIVERSITY_BONUS to module constants
