+++ progress.txt
@@ -1696,3 +1696,11 @@
   signature — a behavior-risk restructure explicitly out of scope for this
   mechanical split — so no core/services/dna/book_sync.py was created.
 - manage.py check clean; test_tasks_integration (7 tests) OK.
+
+## 2026-07-04 17:04 — US-053
+- Moved assign_reader_type verbatim into core/services/dna/reader_type.py.
+  Promoted its magic diversity numbers to module constants:
+  DIVERSITY_THRESHOLD = 10, DIVERSITY_BONUS = 15 (previously function-local).
+- dna/__init__ re-exports assign_reader_type from .reader_type; no patchers
+  target it (only direct imports via the shim, unchanged until US-055).
+- manage.py check clean; test_math_accuracy + test_tasks_unit (59 tests) OK.
